lossless-claw-c02: extract heartbeat pruning helpers
